重新适配了公司详情里的多边形图、文章 阅、评 逻辑调整

+++ SinoNews/SinoNews/Class/RankingList/Model/RankingModel.h
@@ -14,5 +14,5 @@
 @property (nonatomic,strong) NSString *rankingLogo;
 @property (nonatomic,strong) NSString *rankingName; //名称
 @property (nonatomic,strong) NSString *updateTime;
-@property (nonatomic,assign) NSInteger score;       //评分
+@property (nonatomic,assign) double score;       //评分
 @end
